requirements.txt
Flask==3.0.0
flask-cors==4.0.0
qrcode[pil]==7.4.2
Pillow>=10.0.0
